renamed everything from websocket to socketIo
